cv/device-decoding/requirements.txt
opencv-python
depthai==2.16.0.0
depthai-sdk>=1.2.0
numpy
pyserial
pynmea2
numpy
